+++ opennms-services/src/main/java/org/opennms/netmgt/rtc/utils/LegacyEuiLevelMapper.java
@@ -103,19 +103,19 @@ public class LegacyEuiLevelMapper {
             levelCat.setCatvalue(m_dataMgr.getValue(rtcCat, curTime, rWindow));
 
             // nodes in this category
-            for (int nodeID : m_dataMgr.getNodes(rtcCat)) {
+            for (Integer nodeID : m_dataMgr.getNodes(rtcCat)) {
 
                 Node levelNode = new Node();
-                levelNode.setNodeid(nodeID);
+                levelNode.setNodeid(nodeID.longValue());
 
                 // value for this node for this category
                 levelNode.setNodevalue(m_dataMgr.getValue(nodeID, rtcCat, curTime, rWindow));
 
                 // node service count
-                levelNode.setNodesvccount(m_dataMgr.getServiceCount(nodeID, rtcCat));
+                levelNode.setNodesvccount(Long.valueOf(m_dataMgr.getServiceCount(nodeID, rtcCat)));
 
                 // node service down count
-                levelNode.setNodesvcdowncount(m_dataMgr.getServiceDownCount(nodeID, rtcCat));
+                levelNode.setNodesvcdowncount(Long.valueOf(m_dataMgr.getServiceDownCount(nodeID, rtcCat)));
 
                 // add the node
                 levelCat.addNode(levelNode);
